jak3: fix `was-pre-game` crash and `egg-spider` bbush (#4000)
Fixes #3997: The script for the orb that is spawned when getting a medal
in the satellite game was using the wrong entity name.
Fixes #3998: The `spider-manager` process was invalid for one frame,
causing the score to be set to the value of `#f`, winning the game
instantly.

Also doubled the PC port texture count, which should hopefully fix the
crash that happens when playing the game all the way through to the end
of the Destroy Dark Eco Tanks mission in one sitting.